pyproject.toml
[project]
name = "langchain-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.192.0",
    "googlesearch-python>=1.3.0",
    "html-to-markdown>=2.28.2",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-mcp-adapters>=0.2.1",
    "langgraph-cli[inmem]>=0.4.12",
    "langgraph[inmem]>=1.0.10",
    "pygoogle-simple>=0.2.3",
    "scikit-learn>=1.8.0",
    "search-web>=0.1.6",
    "sentence-transformers>=5.2.3",
    "tavily>=1.1.0",
    "tavily-python>=0.7.22",
    "urlpath>=2.0.0",
]
